# [Merged by Bors] - Nested spawns on scope