--- a/crates/bevy_ecs/src/schedule/executor_parallel.rs
+++ b/crates/bevy_ecs/src/schedule/executor_parallel.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use crate::{
     archetype::ArchetypeComponentId,
     query::Access,
@@ -125,7 +127,7 @@ impl ParallelSystemExecutor for ParallelExecutor {
         }
 
         ComputeTaskPool::init(TaskPool::default).scope(|scope| {
-            self.prepare_systems(scope, systems, world);
+            self.prepare_systems(scope.clone(), systems, world);
             if self.should_run.count_ones(..) == 0 {
                 return;
             }
@@ -166,7 +168,7 @@ impl ParallelExecutor {
     /// queues systems with no dependencies to run (or skip) at next opportunity.
     fn prepare_systems<'scope>(
         &mut self,
-        scope: &mut Scope<'scope, ()>,
+        scope: Arc<Scope<'scope, ()>>,
         systems: &'scope mut [ParallelSystemContainer],
         world: &'scope World,
     ) {