--- a/crates/bevy_tasks/Cargo.toml
+++ b/crates/bevy_tasks/Cargo.toml
@@ -15,6 +15,7 @@ async-executor = "1.3.0"
 async-channel = "1.4.2"
 num_cpus = "1"
 once_cell = "1.7"
+concurrent-queue = "1.2.2"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen-futures = "0.4"