--- a/crates/bevy_tasks/src/task_pool.rs
+++ b/crates/bevy_tasks/src/task_pool.rs
@@ -256,7 +256,7 @@ impl TaskPool {
                     self.executor.try_tick();
                     task_scope_executor.try_tick();
 
-                    future::yield_now().await
+                    future::yield_now().await;
                 }
             };