@hymm 2022-04-22
## Perf

Perf tests are a bit of a mixed bag currently. Faster on some tests, but slower on others. This plays out with the examples too where many_cubes is about 2% slower and many_lights is about 2% faster.

![image](https://user-images.githubusercontent.com/2180432/164610944-a8c399d8-525f-4220-9ddd-8e7cefc28e20.png)

![image](https://user-images.githubusercontent.com/2180432/164610992-707627bd-d091-4c14-9e2d-2a8cce8e6308.png)

![image](https://user-images.githubusercontent.com/2180432/164611026-65ddc4f8-ace1-413a-b9a6-b9a9dabca0af.png)