@hymm 2022-04-22
### Lifetimes

The current PR allows the user to get a `&'static Scope`. This would be unsound if the user was able to use this reference after `Scope` gets dropped. I added a compile fail test for having the &Scope reference passed to a 'static thread and tried some other things, but wasn't able to pass the reference to anything other than spawned tasks on the scope. The invariance of the `'scope` lifetime seems to keep the reference properly bound to the closure.

Ideally instead of passing an `&'env Scope` we would pass a `&'scope Scope`, but the `par_for_each` implementation on query errors with the `&'scope` lifetime. This is caused by this bug in rust: https://github.com/rust-lang/rust/issues/95527, and will be fixed when the `nll` nightly feature gets stabilized.  We should change the lifetimes when `nll` is stabilized. https://github.com/rust-lang/rust/issues/43234