@hymm 2022-04-23
Done messing around with this now barring any random ideas I might get for improving the performance.