@hymm 2022-04-27
Added a span to record the scope. Screenshot is the comparison between main and this branch with many cubes example.

<img width="728" alt="Screenshot 2022-04-27 122622" src="https://user-images.githubusercontent.com/2180432/165612279-c0570e3a-f2b1-4ae3-9e60-fac39bc04d0d.png">